Isolation of search code, calling services.
