+++ api/src/main/java/marquez/MarquezContext.java
@@ -43,19 +43,7 @@ import marquez.db.SourceDao;
 import marquez.db.TagDao;
 import marquez.graphql.GraphqlSchemaBuilder;
 import marquez.graphql.MarquezGraphqlServletBuilder;
-import marquez.service.ColumnLineageService;
-import marquez.service.DatasetFieldService;
-import marquez.service.DatasetService;
-import marquez.service.DatasetVersionService;
-import marquez.service.JobService;
-import marquez.service.LineageService;
-import marquez.service.NamespaceService;
-import marquez.service.OpenLineageService;
-import marquez.service.RunService;
-import marquez.service.RunTransitionListener;
-import marquez.service.ServiceFactory;
-import marquez.service.SourceService;
-import marquez.service.TagService;
+import marquez.service.*;
 import marquez.service.models.Tag;
 import org.jdbi.v3.core.Jdbi;
 import org.opensearch.client.opensearch.OpenSearchClient;
@@ -90,6 +78,7 @@ public final class MarquezContext {
   @Getter private final OpenLineageService openLineageService;
   @Getter private final LineageService lineageService;
   @Getter private final ColumnLineageService columnLineageService;
+  @Getter private final SearchService searchService;
   @Getter private final NamespaceResource namespaceResource;
   @Getter private final SourceResource sourceResource;
   @Getter private final DatasetResource datasetResource;
@@ -145,6 +134,7 @@ public final class MarquezContext {
     this.openLineageService = new OpenLineageService(baseDao, runService);
     this.lineageService = new LineageService(lineageDao, jobDao);
     this.columnLineageService = new ColumnLineageService(columnLineageDao, datasetFieldDao);
+    this.searchService = new SearchService(openSearchClient);
     this.jdbiException = new JdbiExceptionExceptionMapper();
     this.jsonException = new JsonProcessingExceptionMapper();
     final ServiceFactory serviceFactory =
@@ -155,6 +145,7 @@ public final class MarquezContext {
             .namespaceService(namespaceService)
             .tagService(tagService)
             .openLineageService(openLineageService)
+            .searchService(searchService)
             .sourceService(sourceService)
             .lineageService(lineageService)
             .columnLineageService(columnLineageService)
@@ -168,8 +159,8 @@ public final class MarquezContext {
     this.jobResource = new JobResource(serviceFactory, jobVersionDao, jobFacetsDao, runFacetsDao);
     this.tagResource = new TagResource(serviceFactory);
     this.openLineageResource =
-        new OpenLineageResource(serviceFactory, openSearchClient, openLineageDao);
-    this.searchResource = new SearchResource(searchDao, openSearchClient);
+        new OpenLineageResource(serviceFactory, openLineageDao);
+    this.searchResource = new SearchResource(serviceFactory, searchDao, openSearchClient);
 
     this.resources =
         ImmutableList.of(
