+++ api/src/main/java/marquez/service/ServiceFactory.java
@@ -23,4 +23,5 @@ public class ServiceFactory {
   @NonNull DatasetFieldService datasetFieldService;
   @NonNull LineageService lineageService;
   @NonNull ColumnLineageService columnLineageService;
+  @NonNull SearchService searchService;
 }
